# docs/swagger.yaml
basePath: /api
definitions:
...
  controllers.DocumentRequest:
    properties:
      agama:
        example: Islam
        type: string
      alamat:
        example: JL. MERDEKA NO. 10, JAKARTA
        type: string
      items:
        items:
          properties:
            deskripsi:
              example: 'NIK: 3171234567890001'
              type: string
            nama_barang:
              example: KTP
              type: string
          required:
          - nama_barang
          type: object
        minItems: 1
        type: array
      jenis_kelamin:
        enum:
        - Laki-laki
        - Perempuan
        type: string
      lokasi_hilang:
        example: Sekitar Pasar Senen
        type: string
      nama_lengkap:
        example: BUDI SANTOSO
        type: string
      pejabat_persetuju_id:
        example: 1
        type: integer
      pekerjaan:
        example: Karyawan Swasta
        type: string
      petugas_pelapor_id:
        example: 2
        type: integer
      tanggal_lahir:
        example: "1990-01-15"
        type: string
      tempat_lahir:
        example: JAKARTA
        type: string
    required:
    - agama
    - alamat
    - items
    - jenis_kelamin
    - lokasi_hilang
    - nama_lengkap
    - pejabat_persetuju_id
    - pekerjaan
    - petugas_pelapor_id
    - tanggal_lahir
    - tempat_lahir
    type: object
...
  dto.AppConfig:
    properties:
      archive_duration_days:
        type: integer
      backup_path:
        type: string
      format_nomor_surat:
        type: string
      is_setup_complete:
        type: boolean
      kop_baris_1:
        type: string
      kop_baris_2:
        type: string
      kop_baris_3:
        type: string
      nama_kantor:
        type: string
      nomor_surat_terakhir:
        type: string
      tempat_surat:
        type: string
      zona_waktu:
        type: string
    type: object
...
  models.LostDocument:
    properties:
      created_at:
        type: string
      id:
        type: integer
      last_updated_by:
        $ref: '#/definitions/models.User'
      last_updated_by_id:
        description: LastUpdatedByID adalah pengguna yang login dan melakukan aksi Update
          terakhir
        type: integer
      lokasi_hilang:
        type: string
      lost_items:
        items:
          $ref: '#/definitions/models.LostItem'
        type: array
      nomor_surat:
        type: string
      operator:
        $ref: '#/definitions/models.User'
      operator_id:
        description: |-
          --- FIELD BARU UNTUK AUDIT ---
          Operator adalah pengguna yang login dan melakukan aksi Create
        type: integer
      pejabat_persetuju:
        $ref: '#/definitions/models.User'
      pejabat_persetuju_id:
        type: integer
      petugas_pelapor:
        $ref: '#/definitions/models.User'
      petugas_pelapor_id:
        description: Petugas yang namanya akan tercetak di surat
        type: integer
      resident:
        $ref: '#/definitions/models.Resident'
      resident_id:
        type: integer
      status:
        type: string
      tanggal_laporan:
        type: string
      tanggal_persetujuan:
        type: string
      updated_at:
        type: string
    type: object
  models.LostItem:
    properties:
      deskripsi:
        type: string
      id:
        type: integer
      lost_document_id:
        type: integer
      nama_barang:
        type: string
    type: object
  models.Resident:
    properties:
      agama:
        type: string
      alamat:
        type: string
      created_at:
        type: string
      id:
        type: integer
      jenis_kelamin:
        type: string
      nama_lengkap:
        type: string
      nik:
        type: string
      pekerjaan:
        type: string
      tanggal_lahir:
        type: string
      tempat_lahir:
        type: string
      updated_at:
        type: string
    type: object
  models.User:
    properties:
      created_at:
        type: string
      id:
        type: integer
      jabatan:
        description: KANIT SPKT, ANGGOTA JAGA REGU
        type: string
      nama_lengkap:
        type: string
      nrp:
        type: string
      pangkat:
        type: string
      peran:
        description: SUPER_ADMIN, OPERATOR
        type: string
      regu:
        description: I, II, III
        type: string
      updated_at:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: yusuf.aar@gmail.com
    name: MUHAMMAD YUSUF ABDURROHMAN
    url: https://github.com/dope-s/simdokpol-go
  description: Ini adalah dokumentasi API untuk aplikasi Sistem Informasi Manajemen
    Dokumen Kepolisian.
  license:
    name: MIT License
    url: https://github.com/dope-s/simdokpol-go/blob/main/LICENSE
  termsOfService: http://swagger.io/terms/
  title: SIMDOKPOL API
  version: "1.0"
paths:
...
  /documents:
    get:
      description: Mengambil daftar semua surat keterangan hilang, bisa difilter berdasarkan
        status (aktif/arsip) dan query pencarian.
      parameters:
      - description: Kata Kunci Pencarian (No. Surat / Nama)
        in: query
        name: q
        type: string
      - default: active
        description: Filter status dokumen
        enum:
        - active
        - archived
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.LostDocument'
            type: array
        "500":
          description: 'Error: Terjadi kesalahan pada server'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Mendapatkan Semua Dokumen
      tags:
      - Documents
    post:
      consumes:
      - application/json
      description: Membuat surat keterangan hilang baru.
      parameters:
      - description: Data Dokumen Baru
        in: body
        name: document
        required: true
        schema:
          $ref: '#/definitions/controllers.DocumentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.LostDocument'
        "400":
          description: 'Error: Input tidak valid'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'Error: Gagal membuat dokumen'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Membuat Dokumen Baru
      tags:
      - Documents
  /documents/export:
    get:
      description: Mengunduh daftar dokumen (aktif atau arsip) sebagai file .xlsx.
        Hanya bisa diakses oleh Super Admin.
      parameters:
      - description: Kata Kunci Pencarian (No. Surat / Nama)
        in: query
        name: q
        type: string
      - default: active
        description: Filter status dokumen
        enum:
        - active
        - archived
        in: query
        name: status
        type: string
      produces:
      - application/vnd.openxmlformats-officedocument.spreadsheetml.sheet
      responses:
        "200":
          description: File Excel (.xlsx)
          schema:
            type: file
        "500":
          description: 'Error: Gagal membuat file ekspor'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Ekspor Dokumen ke Excel
      tags:
      - Documents
  /documents/{id}:
    delete:
      description: Menghapus (soft delete) sebuah surat keterangan hilang. Hanya bisa
        diakses oleh Super Admin atau operator yang membuatnya.
      produces:
      - application/json
      parameters:
      - description: ID Dokumen
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: Pesan Sukses
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 'Error: ID tidak valid'
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 'Error: Akses ditolak'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'Error: Gagal menghapus dokumen'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Menghapus Dokumen
      tags:
      - Documents
    get:
      description: Mengambil detail satu surat keterangan hilang berdasarkan ID-nya.
        Hanya bisa diakses oleh Super Admin atau operator yang membuat dokumen tersebut.
      produces:
      - application/json
      parameters:
      - description: ID Dokumen
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.LostDocument'
        "400":
          description: 'Error: ID tidak valid'
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 'Error: Akses ditolak'
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 'Error: Dokumen tidak ditemukan'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Mendapatkan Dokumen Berdasarkan ID
      tags:
      - Documents
    put:
      consumes:
      - application/json
      description: Memperbarui data sebuah surat keterangan hilang. Hanya bisa diakses
        oleh Super Admin atau operator yang membuatnya.
      parameters:
      - description: ID Dokumen
        in: path
        name: id
        required: true
        type: integer
      - description: Data Dokumen yang Diperbarui
        in: body
        name: document
        required: true
        schema:
          $ref: '#/definitions/controllers.DocumentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.LostDocument'
        "400":
          description: 'Error: Input tidak valid'
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 'Error: Akses ditolak'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'Error: Gagal memperbarui dokumen'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Memperbarui Dokumen
      tags:
      - Documents
  /documents/{id}/pdf:
    get:
      description: Mengunduh satu surat keterangan hilang sebagai file .pdf yang presisi.
      produces:
      - application/pdf
      parameters:
      - description: ID Dokumen
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: File PDF (.pdf)
          schema:
            type: file
        "400":
          description: 'Error: ID tidak valid'
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 'Error: Akses ditolak'
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 'Error: Dokumen tidak ditemukan'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'Error: Gagal membuat file PDF'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Download Dokumen sebagai PDF
      tags:
      - Documents
...
  /search:
    get:
      description: Mencari dokumen (aktif dan arsip) berdasarkan Nomor Surat atau Nama
        Pemohon.
      parameters:
      - description: Kata Kunci Pencarian
        in: query
        name: q
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.LostDocument'
            type: array
        "500":
          description: 'Error: Terjadi kesalahan pada server'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Pencarian Dokumen Global
      tags:
      - Documents
  /settings:
    get:
      description: Mengambil semua data konfigurasi sistem yang sedang aktif. Hanya
        bisa diakses oleh Super Admin.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AppConfig'
        "500":
          description: 'Error: Gagal mengambil data pengaturan'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Mendapatkan Semua Pengaturan Sistem
      tags:
      - Settings
    put:
      consumes:
      - application/json
      description: Menyimpan satu atau lebih data konfigurasi sistem. Hanya bisa diakses
        oleh Super Admin.
      parameters:
      - description: Data Pengaturan Baru
        in: body
        name: settings
        required: true
        schema:
          $ref: '#/definitions/dto.AppConfig'
      produces:
      - application/json
      responses:
        "200":
          description: Pesan Sukses
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 'Error: Format data tidak valid'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'Error: Gagal menyimpan pengaturan'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Memperbarui Pengaturan Sistem
      tags:
      - Settings
...
securityDefinitions:
  BearerAuth:
    description: Masukkan token JWT Anda dengan format 'Bearer {token}'.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"